Add tests and travis configuration
Adds, some mocha unit test, this is pretty basic for now, but allows for
expansion of more tests later on. Also includes travis-ci configuration,
this might have some issues due to the dependency on PowerShell.

Last but not least, this adds somes badges til to readme about the
build, version and dependencies.